+ cloth resizing

+++ OpenGL4/OpenGL4/Scene_Clothsim.h
@@ -57,9 +57,9 @@ private:
 	
 	bool m_DebugDraw = false;
 	bool m_WireFrame = false;
-	float m_ClothLength = 20.0f;
-	float m_ClothWidth = 20.0f;
-	int m_HookCount = (int)m_ClothWidth;
+	int m_ClothLength = 20;
+	int m_ClothWidth = 20;
+	int m_HookCount = m_ClothWidth;
 	float m_HookDistance{1.0f};
 	float m_Stiffness = 500.0f;
 	int m_SelectedMouseMode{};
